Removed old-style casts Added support for XMT compilation
git-svn-id: https://svn.vital-it.ch/svn/hp2c/trunk/Codeml_Variants/Fastcodeml@3716 95c5a404-1f88-0410-a6b0-c3f062b6f34a

+++ Forest.cpp
@@ -478,7 +478,7 @@ void Forest::groupByDependency(bool aForceSerial)
 	{
 		std::vector<unsigned int> v(mNumSites);
 
-		for(unsigned int k=0; k < (unsigned int)mNumSites; ++k) v[k] = (unsigned int)mNumSites-k-1; // Remember: prior (could) point to subsequent
+		for(size_t k=0; k < mNumSites; ++k) v[k] = static_cast<unsigned int>(mNumSites-k-1); // Remember: prior (could) point to subsequent
 
 		class_dependencies.push_back(v);
 	}
